# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/factors/sources": {
      "get": {
        "operationId": "FactorsController_sources",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User manages factor governance; other roles can read factor catalogues.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "factors"
        ]
      },
      "post": {
        "operationId": "FactorsController_createSource",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User manages factor governance; other roles can read factor catalogues.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FactorSourceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Standard response format."
          },
          "201": {
            "description": ""
          }
        },
        "tags": [
          "factors"
        ]
      }
    },
    "/api/factors/sources/{id}": {
      "patch": {
        "operationId": "FactorsController_updateSource",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User manages factor governance; other roles can read factor catalogues.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFactorSourceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "factors"
        ]
      },
      "put": {
        "operationId": "FactorsController_replaceSource",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User manages factor governance; other roles can read factor catalogues.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFactorSourceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "factors"
        ]
      },
      "delete": {
        "operationId": "FactorsController_removeSource",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User manages factor governance; other roles can read factor catalogues.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "factors"
        ]
      }
    },
    "/api/factors/versions": {
      "get": {
        "operationId": "FactorsController_versions",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User manages factor governance; other roles can read factor catalogues.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "factors"
        ]
      },
      "post": {
        "operationId": "FactorsController_createVersion",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User manages factor governance; other roles can read factor catalogues.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FactorVersionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Standard response format."
          },
          "201": {
            "description": ""
          }
        },
        "tags": [
          "factors"
        ]
      }
    },
    "/api/factors/versions/{id}": {
      "patch": {
        "operationId": "FactorsController_updateVersion",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User manages factor governance; other roles can read factor catalogues.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFactorVersionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "factors"
        ]
      },
      "put": {
        "operationId": "FactorsController_replaceVersion",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User manages factor governance; other roles can read factor catalogues.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFactorVersionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "factors"
        ]
      },
      "delete": {
        "operationId": "FactorsController_removeVersion",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User manages factor governance; other roles can read factor catalogues.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "factors"
        ]
      }
    },
    "/api/factors": {
      "get": {
        "operationId": "FactorsController_list",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User manages factor governance; other roles can read factor catalogues.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "factors"
        ]
      },
      "post": {
        "operationId": "FactorsController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User manages factor governance; other roles can read factor catalogues.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmissionFactorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Standard response format."
          },
          "201": {
            "description": ""
          }
        },
        "tags": [
          "factors"
        ]
      }
    },
    "/api/factors/{id}": {
      "patch": {
        "operationId": "FactorsController_update",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User manages factor governance; other roles can read factor catalogues.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmissionFactorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "factors"
        ]
      },
      "put": {
        "operationId": "FactorsController_replace",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User manages factor governance; other roles can read factor catalogues.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmissionFactorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "factors"
        ]
      },
      "delete": {
        "operationId": "FactorsController_remove",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User manages factor governance; other roles can read factor catalogues.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "factors"
        ]
      }
    },
...
  },
  "info": {
    "title": "Rorizon Review-4 API",
    "description": "NestJS in-memory backend with RBAC role headers for Rorizon workflows.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "Role for RBAC authorization. Allowed values: Super User, COO, Manager, Analyst."
      }
    },
    "schemas": {
...
      "FactorSourceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "efs-rorizon-demo"
          },
          "name": {
            "type": "string",
            "example": "Rorizon Demo Factors"
          },
          "publisher": {
            "type": "string",
            "example": "Platform Governance"
          },
          "region": {
            "type": "string",
            "example": "IN"
          },
          "datasetName": {
            "type": "string",
            "example": "Published conversion factor dataset"
          },
          "datasetUrl": {
            "type": "string",
            "example": "https://example.org/dataset"
          },
          "acquisitionMethod": {
            "type": "string",
            "example": "Published Dataset",
            "enum": [
              "Manual Upload",
              "Published Dataset",
              "API",
              "Other"
            ]
          },
          "license": {
            "type": "string",
            "example": "Open Government License"
          },
          "notes": {
            "type": "string",
            "example": "Imported by Super User after dataset review."
          },
          "active": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "name",
          "publisher",
          "region"
        ]
      },
      "UpdateFactorSourceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "efs-rorizon-demo"
          },
          "name": {
            "type": "string",
            "example": "Rorizon Demo Factors"
          },
          "publisher": {
            "type": "string",
            "example": "Platform Governance"
          },
          "region": {
            "type": "string",
            "example": "IN"
          },
          "datasetName": {
            "type": "string",
            "example": "Published conversion factor dataset"
          },
          "datasetUrl": {
            "type": "string",
            "example": "https://example.org/dataset"
          },
          "acquisitionMethod": {
            "type": "string",
            "example": "Published Dataset",
            "enum": [
              "Manual Upload",
              "Published Dataset",
              "API",
              "Other"
            ]
          },
          "license": {
            "type": "string",
            "example": "Open Government License"
          },
          "notes": {
            "type": "string",
            "example": "Imported by Super User after dataset review."
          },
          "active": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "FactorVersionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "fv-2026-demo"
          },
          "sourceId": {
            "type": "string",
            "example": "efs-rorizon-demo"
          },
          "name": {
            "type": "string",
            "example": "2026 Demo Baseline"
          },
          "datasetName": {
            "type": "string",
            "example": "Government conversion factors"
          },
          "datasetYear": {
            "type": "string",
            "example": "2026"
          },
          "importedAt": {
            "type": "string",
            "example": "2026-04-01T00:00:00.000Z"
          },
          "importedBy": {
            "type": "string",
            "example": "System Admin"
          },
          "effectiveFrom": {
            "type": "string",
            "example": "2026-01-01"
          },
          "effectiveTo": {
            "type": "string",
            "example": "2026-12-31"
          },
          "status": {
            "type": "string",
            "example": "Active",
            "enum": [
              "Draft",
              "Active",
              "Archived",
              "Retired"
            ]
          },
          "locked": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "sourceId",
          "name",
          "effectiveFrom"
        ]
      },
      "UpdateFactorVersionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "fv-2026-demo"
          },
          "sourceId": {
            "type": "string",
            "example": "efs-rorizon-demo"
          },
          "name": {
            "type": "string",
            "example": "2026 Demo Baseline"
          },
          "datasetName": {
            "type": "string",
            "example": "Government conversion factors"
          },
          "datasetYear": {
            "type": "string",
            "example": "2026"
          },
          "importedAt": {
            "type": "string",
            "example": "2026-04-01T00:00:00.000Z"
          },
          "importedBy": {
            "type": "string",
            "example": "System Admin"
          },
          "effectiveFrom": {
            "type": "string",
            "example": "2026-01-01"
          },
          "effectiveTo": {
            "type": "string",
            "example": "2026-12-31"
          },
          "status": {
            "type": "string",
            "example": "Active",
            "enum": [
              "Draft",
              "Active",
              "Archived",
              "Retired"
            ]
          },
          "locked": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "EmissionFactorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "ef-diesel-liter-2026"
          },
          "sourceId": {
            "type": "string",
            "example": "efs-rorizon-demo"
          },
          "versionId": {
            "type": "string",
            "example": "fv-2026-demo"
          },
          "resourceTypeId": {
            "type": "string",
            "example": "rt-diesel"
          },
          "unitId": {
            "type": "string",
            "example": "unit-liter"
          },
          "factor": {
            "type": "number",
            "example": 2.5
          },
          "factorUnit": {
            "type": "string",
            "example": "kgCO2e/L"
          },
          "geography": {
            "type": "string",
            "example": "IN"
          },
          "scope": {
            "type": "string",
            "example": "Scope 1"
          },
          "activityBoundary": {
            "type": "string",
            "example": "Combustion, tank-to-wheel"
          },
          "uncertainty": {
            "type": "string",
            "example": "Medium"
          },
          "notes": {
            "type": "string",
            "example": "Use only for stationary diesel combustion."
          },
          "validFrom": {
            "type": "string",
            "example": "2026-01-01"
          },
          "validTo": {
            "type": "string",
            "example": "2026-12-31"
          },
          "active": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "sourceId",
          "versionId",
          "resourceTypeId",
          "unitId",
          "factor",
          "factorUnit"
        ]
      },
      "UpdateEmissionFactorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "ef-diesel-liter-2026"
          },
          "sourceId": {
            "type": "string",
            "example": "efs-rorizon-demo"
          },
          "versionId": {
            "type": "string",
            "example": "fv-2026-demo"
          },
          "resourceTypeId": {
            "type": "string",
            "example": "rt-diesel"
          },
          "unitId": {
            "type": "string",
            "example": "unit-liter"
          },
          "factor": {
            "type": "number",
            "example": 2.5
          },
          "factorUnit": {
            "type": "string",
            "example": "kgCO2e/L"
          },
          "geography": {
            "type": "string",
            "example": "IN"
          },
          "scope": {
            "type": "string",
            "example": "Scope 1"
          },
          "activityBoundary": {
            "type": "string",
            "example": "Combustion, tank-to-wheel"
          },
          "uncertainty": {
            "type": "string",
            "example": "Medium"
          },
          "notes": {
            "type": "string",
            "example": "Use only for stationary diesel combustion."
          },
          "validFrom": {
            "type": "string",
            "example": "2026-01-01"
          },
          "validTo": {
            "type": "string",
            "example": "2026-12-31"
          },
          "active": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
    }
  }
}